+++ hutool-db/src/main/java/cn/hutool/db/dialect/DialectFactory.java
@@ -54,6 +54,8 @@ public class DialectFactory {
 	public static final String DRIVER_KINGBASE8 = "com.kingbase8.Driver";
 	/** JDBC 驱动 Ignite thin */
 	public static final String DRIVER_IGNITE_THIN = "org.apache.ignite.IgniteJdbcThinDriver";
+	/** JDBC 驱动 ClickHouse */
+	public static final String DRIVER_CLICK_HOUSE = "ru.yandex.clickhouse.ClickHouseDriver";
 
 	private static final Map<DataSource, Dialect> DIALECT_POOL = new ConcurrentHashMap<>();
 
@@ -143,6 +145,9 @@ public class DialectFactory {
 		} else if (nameContainsProductInfo.contains("ignite")) {
 			// Ignite thin
 			driver = DRIVER_IGNITE_THIN;
+		} else if (nameContainsProductInfo.contains("clickhouse")) {
+			// ClickHouse
+			driver = DRIVER_CLICK_HOUSE;
 		}
 
 		return driver;